backend/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
stripe==2.49.0
vcrpy==4.0.2 # test-only
flake8==3.8.3 # dev-only
black==19.10b0 # dev-only
isort==5.2.0 # dev-only